move invalidate multiple queries to hook, fix lb cache invalidation
